requirements.txt
telethon
yagmail
python-dotenv